Refactor API client into a separate package.
