+++ api/client/client.go
@@ -1,0 +1,616 @@
+/*
+Copyright 2020 Gravitational, Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+// Package client holds the implementation of the Teleport gRPC api client
+package client
+
+import (
+	"compress/gzip"
+	"context"
+	"crypto/tls"
+	"io"
+	"net"
+	"sync/atomic"
+	"time"
+
+	"github.com/gravitational/teleport"
+	"github.com/gravitational/teleport/api/client/proto"
+	"github.com/gravitational/teleport/lib/defaults"
+	"github.com/gravitational/teleport/lib/events"
+	"github.com/gravitational/teleport/lib/jwt"
+	"github.com/gravitational/teleport/lib/services"
+	"golang.org/x/net/http2"
+
+	"github.com/golang/protobuf/ptypes/empty"
+	"github.com/gravitational/trace"
+	"github.com/gravitational/trace/trail"
+	"google.golang.org/grpc"
+	"google.golang.org/grpc/credentials"
+	ggzip "google.golang.org/grpc/encoding/gzip"
+	"google.golang.org/grpc/keepalive"
+)
+
+func init() {
+	// gzip is used for gRPC auditStream compression. SetLevel changes the
+	// compression level, must be called in initialization, and is not thread safe.
+	if err := ggzip.SetLevel(gzip.BestSpeed); err != nil {
+		panic(err)
+	}
+}
+
+// Client is a gRPC Client that connects to a teleport auth server through TLS.
+type Client struct {
+	c Config
+	// grpc is the gRPC client specification for the auth server.
+	grpc proto.AuthServiceClient
+	// conn is a grpc connection to the auth server.
+	conn *grpc.ClientConn
+	// closedFlag is set to indicate that the services are closed.
+	closedFlag int32
+}
+
+// NewClient returns a new auth client that uses mutual TLS authentication and
+// connects to the remote server using the Dialer or Addrs given in Config.
+func NewClient(cfg Config) (*Client, error) {
+	if err := cfg.CheckAndSetDefaults(); err != nil {
+		return nil, trace.Wrap(err)
+	}
+
+	c := &Client{c: cfg}
+	dialer := grpc.WithContextDialer(func(ctx context.Context, addr string) (net.Conn, error) {
+		if c.isClosed() {
+			return nil, trace.ConnectionProblem(nil, "client is closed")
+		}
+		conn, err := c.c.Dialer.DialContext(ctx, "tcp", addr)
+		if err != nil {
+			return nil, trace.ConnectionProblem(err, "failed to dial")
+		}
+		return conn, nil
+	})
+
+	var err error
+	if c.conn, err = grpc.Dial(teleport.APIDomain,
+		dialer,
+		grpc.WithTransportCredentials(credentials.NewTLS(c.c.TLS)),
+		grpc.WithKeepaliveParams(keepalive.ClientParameters{
+			Time:                c.c.KeepAlivePeriod,
+			Timeout:             c.c.KeepAlivePeriod * time.Duration(c.c.KeepAliveCount),
+			PermitWithoutStream: true,
+		}),
+	); err != nil {
+		return nil, trail.FromGRPC(err)
+	}
+
+	c.grpc = proto.NewAuthServiceClient(c.conn)
+	return c, nil
+}
+
+// Config contains configuration of the client
+type Config struct {
+	// Addrs is a list of teleport auth/proxy server addresses to dial
+	Addrs []string
+	// Dialer is a custom dialer that is used instead of Addrs when provided
+	Dialer ContextDialer
+	// DialTimeout defines how long to attempt dialing before timing out
+	DialTimeout time.Duration
+	// KeepAlivePeriod defines period between keep alives
+	KeepAlivePeriod time.Duration
+	// KeepAliveCount specifies the amount of missed keep alives
+	// to wait for before declaring the connection as broken
+	KeepAliveCount int
+	// TLS is the client's TLS config
+	TLS *tls.Config
+}
+
+// CheckAndSetDefaults checks and sets default config values
+func (c *Config) CheckAndSetDefaults() error {
+	if len(c.Addrs) == 0 && c.Dialer == nil {
+		return trace.BadParameter("set parameter Addrs or Dialer")
+	}
+	if c.TLS == nil {
+		return trace.BadParameter("missing parameter TLS")
+	}
+	c.TLS = c.TLS.Clone()
+	if c.KeepAlivePeriod == 0 {
+		c.KeepAlivePeriod = defaults.ServerKeepAliveTTL
+	}
+	if c.KeepAliveCount == 0 {
+		c.KeepAliveCount = defaults.KeepAliveCountMax
+	}
+	if c.DialTimeout == 0 {
+		c.DialTimeout = defaults.DefaultDialTimeout
+	}
+	if c.Dialer == nil {
+		var err error
+		if c.Dialer, err = NewAddrDialer(c.Addrs, c.KeepAlivePeriod, c.DialTimeout); err != nil {
+			return trace.Wrap(err)
+		}
+	}
+	c.TLS.NextProtos = []string{http2.NextProtoTLS}
+	if c.TLS.ServerName == "" {
+		c.TLS.ServerName = teleport.APIDomain
+	}
+
+	return nil
+}
+
+// Dialer returns the client's connection dialer
+func (c *Client) Dialer() ContextDialer {
+	return c.c.Dialer
+}
+
+// Close closes the Client connection to the auth server
+func (c *Client) Close() error {
+	if !c.setClosed() {
+		return nil
+	}
+	if c.conn != nil {
+		err := c.conn.Close()
+		c.conn = nil
+		return trace.Wrap(err)
+	}
+	return nil
+}
+
+func (c *Client) isClosed() bool {
+	return atomic.LoadInt32(&c.closedFlag) == 1
+}
+
+// setClosed marks the client to closed and returns true
+// if the client was successfully marked as closed.
+func (c *Client) setClosed() bool {
+	return atomic.CompareAndSwapInt32(&c.closedFlag, 0, 1)
+}
+
+// Ping gets basic info about the auth server.
+func (c *Client) Ping(ctx context.Context) (proto.PingResponse, error) {
+	rsp, err := c.grpc.Ping(ctx, &proto.PingRequest{})
+	if err != nil {
+		return proto.PingResponse{}, trail.FromGRPC(err)
+	}
+	return *rsp, nil
+}
+
+// UpsertNode is used by SSH servers to report their presence
+// to the auth servers in form of heartbeat expiring after ttl period.
+func (c *Client) UpsertNode(s services.Server) (*services.KeepAlive, error) {
+	if s.GetNamespace() == "" {
+		return nil, trace.BadParameter("missing node namespace")
+	}
+	protoServer, ok := s.(*services.ServerV2)
+	if !ok {
+		return nil, trace.BadParameter("unsupported client")
+	}
+	keepAlive, err := c.grpc.UpsertNode(context.TODO(), protoServer)
+	if err != nil {
+		return nil, trail.FromGRPC(err)
+	}
+	return keepAlive, nil
+}
+
+// UpdateRemoteCluster updates remote cluster from the specified value.
+func (c *Client) UpdateRemoteCluster(ctx context.Context, rc services.RemoteCluster) error {
+	rcV3, ok := rc.(*services.RemoteClusterV3)
+	if !ok {
+		return trace.BadParameter("unsupported remote cluster type %T", rcV3)
+	}
+
+	_, err := c.grpc.UpdateRemoteCluster(ctx, rcV3)
+	return trail.FromGRPC(err)
+}
+
+// CreateUser creates a new user from the specified descriptor.
+func (c *Client) CreateUser(ctx context.Context, user services.User) error {
+	userV2, ok := user.(*services.UserV2)
+	if !ok {
+		return trace.BadParameter("unsupported user type %T", user)
+	}
+
+	_, err := c.grpc.CreateUser(ctx, userV2)
+	return trail.FromGRPC(err)
+}
+
+// UpdateUser updates an existing user in a backend.
+func (c *Client) UpdateUser(ctx context.Context, user services.User) error {
+	userV2, ok := user.(*services.UserV2)
+	if !ok {
+		return trace.BadParameter("unsupported user type %T", user)
+	}
+
+	_, err := c.grpc.UpdateUser(ctx, userV2)
+	return trail.FromGRPC(err)
+}
+
+// GetUser returns a list of usernames registered in the system.
+// withSecrets controls whether authentication details are returned.
+func (c *Client) GetUser(name string, withSecrets bool) (services.User, error) {
+	if name == "" {
+		return nil, trace.BadParameter("missing username")
+	}
+	user, err := c.grpc.GetUser(context.TODO(), &proto.GetUserRequest{
+		Name:        name,
+		WithSecrets: withSecrets,
+	})
+	if err != nil {
+		return nil, trail.FromGRPC(err)
+	}
+	return user, nil
+}
+
+// GetUsers returns a list of users.
+// withSecrets controls whether authentication details are returned.
+func (c *Client) GetUsers(withSecrets bool) ([]services.User, error) {
+	stream, err := c.grpc.GetUsers(context.TODO(), &proto.GetUsersRequest{
+		WithSecrets: withSecrets,
+	})
+	if err != nil {
+		return nil, trail.FromGRPC(err)
+	}
+	var users []services.User
+	for {
+		user, err := stream.Recv()
+		if err != nil {
+			if err == io.EOF {
+				break
+			}
+			return nil, trail.FromGRPC(err)
+		}
+		users = append(users, user)
+	}
+	return users, nil
+}
+
+// DeleteUser deletes a user by name.
+func (c *Client) DeleteUser(ctx context.Context, user string) error {
+	req := &proto.DeleteUserRequest{Name: user}
+	_, err := c.grpc.DeleteUser(ctx, req)
+	return trail.FromGRPC(err)
+}
+
+// GenerateUserCerts takes the public key in the OpenSSH `authorized_keys` plain
+// text format, signs it using User Certificate Authority signing key and
+// returns the resulting certificates.
+func (c *Client) GenerateUserCerts(ctx context.Context, req proto.UserCertsRequest) (*proto.Certs, error) {
+	certs, err := c.grpc.GenerateUserCerts(ctx, &req)
+	if err != nil {
+		return nil, trail.FromGRPC(err)
+	}
+	return certs, nil
+}
+
+// EmitAuditEvent sends an auditable event to the auth server.
+func (c *Client) EmitAuditEvent(ctx context.Context, event events.AuditEvent) error {
+	grpcEvent, err := events.ToOneOf(event)
+	if err != nil {
+		return trace.Wrap(err)
+	}
+	_, err = c.grpc.EmitAuditEvent(ctx, grpcEvent)
+	if err != nil {
+		return trail.FromGRPC(err)
+	}
+	return nil
+}
+
+// GetAccessRequests retrieves a list of all access requests matching the provided filter.
+func (c *Client) GetAccessRequests(ctx context.Context, filter services.AccessRequestFilter) ([]services.AccessRequest, error) {
+	rsp, err := c.grpc.GetAccessRequests(ctx, &filter)
+	if err != nil {
+		return nil, trail.FromGRPC(err)
+	}
+	reqs := make([]services.AccessRequest, 0, len(rsp.AccessRequests))
+	for _, req := range rsp.AccessRequests {
+		reqs = append(reqs, req)
+	}
+	return reqs, nil
+}
+
+// CreateAccessRequest registers a new access request with the auth server.
+func (c *Client) CreateAccessRequest(ctx context.Context, req services.AccessRequest) error {
+	r, ok := req.(*services.AccessRequestV3)
+	if !ok {
+		return trace.BadParameter("unexpected access request type %T", req)
+	}
+	_, err := c.grpc.CreateAccessRequest(ctx, r)
+	return trail.FromGRPC(err)
+}
+
+// RotateResetPasswordTokenSecrets rotates secrets for a given tokenID.
+// It gets called every time a user fetches 2nd-factor secrets during registration attempt.
+// This ensures that an attacker that gains the ResetPasswordToken link can not view it,
+// extract the OTP key from the QR code, then allow the user to signup with
+// the same OTP token.
+func (c *Client) RotateResetPasswordTokenSecrets(ctx context.Context, tokenID string) (services.ResetPasswordTokenSecrets, error) {
+	secrets, err := c.grpc.RotateResetPasswordTokenSecrets(ctx, &proto.RotateResetPasswordTokenSecretsRequest{
+		TokenID: tokenID,
+	})
+	if err != nil {
+		return nil, trail.FromGRPC(err)
+	}
+	return secrets, nil
+}
+
+// GetResetPasswordToken returns a ResetPasswordToken for the specified tokenID.
+func (c *Client) GetResetPasswordToken(ctx context.Context, tokenID string) (services.ResetPasswordToken, error) {
+	token, err := c.grpc.GetResetPasswordToken(ctx, &proto.GetResetPasswordTokenRequest{
+		TokenID: tokenID,
+	})
+	if err != nil {
+		return nil, trail.FromGRPC(err)
+	}
+
+	return token, nil
+}
+
+// CreateResetPasswordToken creates reset password token.
+func (c *Client) CreateResetPasswordToken(ctx context.Context, req *proto.CreateResetPasswordTokenRequest) (services.ResetPasswordToken, error) {
+	token, err := c.grpc.CreateResetPasswordToken(ctx, req)
+	if err != nil {
+		return nil, trail.FromGRPC(err)
+	}
+
+	return token, nil
+}
+
+// DeleteAccessRequest deletes an access request.
+func (c *Client) DeleteAccessRequest(ctx context.Context, reqID string) error {
+	_, err := c.grpc.DeleteAccessRequest(ctx, &proto.RequestID{ID: reqID})
+	return trail.FromGRPC(err)
+}
+
+// SetAccessRequestState updates the state of an existing access request.
+func (c *Client) SetAccessRequestState(ctx context.Context, params services.AccessRequestUpdate) error {
+	setter := proto.RequestStateSetter{
+		ID:          params.RequestID,
+		State:       params.State,
+		Reason:      params.Reason,
+		Annotations: params.Annotations,
+		Roles:       params.Roles,
+	}
+	if d := GetDelegator(ctx); d != "" {
+		setter.Delegator = d
+	}
+	_, err := c.grpc.SetAccessRequestState(ctx, &setter)
+	return trail.FromGRPC(err)
+}
+
+// GetPluginData loads all plugin data matching the supplied filter.
+func (c *Client) GetPluginData(ctx context.Context, filter services.PluginDataFilter) ([]services.PluginData, error) {
+	seq, err := c.grpc.GetPluginData(ctx, &filter)
+	if err != nil {
+		return nil, trail.FromGRPC(err)
+	}
+	data := make([]services.PluginData, 0, len(seq.PluginData))
+	for _, d := range seq.PluginData {
+		data = append(data, d)
+	}
+	return data, nil
+}
+
+// UpdatePluginData updates a per-resource PluginData entry.
+func (c *Client) UpdatePluginData(ctx context.Context, params services.PluginDataUpdateParams) error {
+	_, err := c.grpc.UpdatePluginData(ctx, &params)
+	return trail.FromGRPC(err)
+}
+
+// AcquireSemaphore acquires lease with requested resources from semaphore.
+func (c *Client) AcquireSemaphore(ctx context.Context, params services.AcquireSemaphoreRequest) (*services.SemaphoreLease, error) {
+	lease, err := c.grpc.AcquireSemaphore(ctx, &params)
+	if err != nil {
+		return nil, trail.FromGRPC(err)
+	}
+	return lease, nil
+}
+
+// KeepAliveSemaphoreLease updates semaphore lease.
+func (c *Client) KeepAliveSemaphoreLease(ctx context.Context, lease services.SemaphoreLease) error {
+	_, err := c.grpc.KeepAliveSemaphoreLease(ctx, &lease)
+	return trail.FromGRPC(err)
+}
+
+// CancelSemaphoreLease cancels semaphore lease early.
+func (c *Client) CancelSemaphoreLease(ctx context.Context, lease services.SemaphoreLease) error {
+	_, err := c.grpc.CancelSemaphoreLease(ctx, &lease)
+	return trail.FromGRPC(err)
+}
+
+// GetSemaphores returns a list of all semaphores matching the supplied filter.
+func (c *Client) GetSemaphores(ctx context.Context, filter services.SemaphoreFilter) ([]services.Semaphore, error) {
+	rsp, err := c.grpc.GetSemaphores(ctx, &filter)
+	if err != nil {
+		return nil, trail.FromGRPC(err)
+	}
+	sems := make([]services.Semaphore, 0, len(rsp.Semaphores))
+	for _, s := range rsp.Semaphores {
+		sems = append(sems, s)
+	}
+	return sems, nil
+}
+
+// DeleteSemaphore deletes a semaphore matching the supplied filter.
+func (c *Client) DeleteSemaphore(ctx context.Context, filter services.SemaphoreFilter) error {
+	_, err := c.grpc.DeleteSemaphore(ctx, &filter)
+	return trail.FromGRPC(err)
+}
+
+// UpsertKubeService is used by kubernetes services to report their presence
+// to other auth servers in form of hearbeat expiring after ttl period.
+func (c *Client) UpsertKubeService(ctx context.Context, s services.Server) error {
+	server, ok := s.(*services.ServerV2)
+	if !ok {
+		return trace.BadParameter("invalid type %T, expected *services.ServerV2", server)
+	}
+	_, err := c.grpc.UpsertKubeService(ctx, &proto.UpsertKubeServiceRequest{
+		Server: server,
+	})
+	return trace.Wrap(err)
+}
+
+// GetKubeServices returns the list of kubernetes services registered in the
+// cluster.
+func (c *Client) GetKubeServices(ctx context.Context) ([]services.Server, error) {
+	resp, err := c.grpc.GetKubeServices(ctx, &proto.GetKubeServicesRequest{})
+	if err != nil {
+		return nil, trace.Wrap(err)
+	}
+
+	var servers []services.Server
+	for _, server := range resp.GetServers() {
+		servers = append(servers, server)
+	}
+	return servers, nil
+}
+
+// GetAppServers gets all application servers.
+func (c *Client) GetAppServers(ctx context.Context, namespace string, opts ...services.MarshalOption) ([]services.Server, error) {
+	cfg, err := services.CollectOptions(opts)
+	if err != nil {
+		return nil, trace.Wrap(err)
+	}
+
+	resp, err := c.grpc.GetAppServers(ctx, &proto.GetAppServersRequest{
+		Namespace:      namespace,
+		SkipValidation: cfg.SkipValidation,
+	})
+	if err != nil {
+		return nil, trail.FromGRPC(err)
+	}
+
+	var servers []services.Server
+	for _, server := range resp.GetServers() {
+		servers = append(servers, server)
+	}
+
+	return servers, nil
+}
+
+// UpsertAppServer adds an application server.
+func (c *Client) UpsertAppServer(ctx context.Context, server services.Server) (*services.KeepAlive, error) {
+	s, ok := server.(*services.ServerV2)
+	if !ok {
+		return nil, trace.BadParameter("invalid type %T", server)
+	}
+
+	keepAlive, err := c.grpc.UpsertAppServer(ctx, &proto.UpsertAppServerRequest{
+		Server: s,
+	})
+	if err != nil {
+		return nil, trail.FromGRPC(err)
+	}
+	return keepAlive, nil
+}
+
+// DeleteAppServer removes an application server.
+func (c *Client) DeleteAppServer(ctx context.Context, namespace string, name string) error {
+	_, err := c.grpc.DeleteAppServer(ctx, &proto.DeleteAppServerRequest{
+		Namespace: namespace,
+		Name:      name,
+	})
+	return trail.FromGRPC(err)
+}
+
+// DeleteAllAppServers removes all application servers.
+func (c *Client) DeleteAllAppServers(ctx context.Context, namespace string) error {
+	_, err := c.grpc.DeleteAllAppServers(ctx, &proto.DeleteAllAppServersRequest{
+		Namespace: namespace,
+	})
+	return trail.FromGRPC(err)
+}
+
+// GetAppSession gets an application web session.
+func (c *Client) GetAppSession(ctx context.Context, req services.GetAppSessionRequest) (services.WebSession, error) {
+	resp, err := c.grpc.GetAppSession(ctx, &proto.GetAppSessionRequest{
+		SessionID: req.SessionID,
+	})
+	if err != nil {
+		return nil, trail.FromGRPC(err)
+	}
+
+	return resp.GetSession(), nil
+}
+
+// GetAppSessions gets all application web sessions.
+func (c *Client) GetAppSessions(ctx context.Context) ([]services.WebSession, error) {
+	resp, err := c.grpc.GetAppSessions(ctx, &empty.Empty{})
+	if err != nil {
+		return nil, trail.FromGRPC(err)
+	}
+
+	out := make([]services.WebSession, 0, len(resp.GetSessions()))
+	for _, v := range resp.GetSessions() {
+		out = append(out, v)
+	}
+	return out, nil
+}
+
+// CreateAppSession creates an application web session. Application web
+// sessions represent a browser session the client holds.
+func (c *Client) CreateAppSession(ctx context.Context, req services.CreateAppSessionRequest) (services.WebSession, error) {
+	resp, err := c.grpc.CreateAppSession(ctx, &proto.CreateAppSessionRequest{
+		Username:      req.Username,
+		ParentSession: req.ParentSession,
+		PublicAddr:    req.PublicAddr,
+		ClusterName:   req.ClusterName,
+	})
+	if err != nil {
+		return nil, trail.FromGRPC(err)
+	}
+
+	return resp.GetSession(), nil
+}
+
+// DeleteAppSession removes an application web session.
+func (c *Client) DeleteAppSession(ctx context.Context, req services.DeleteAppSessionRequest) error {
+
+	_, err := c.grpc.DeleteAppSession(ctx, &proto.DeleteAppSessionRequest{
+		SessionID: req.SessionID,
+	})
+	return trail.FromGRPC(err)
+}
+
+// DeleteAllAppSessions removes all application web sessions.
+func (c *Client) DeleteAllAppSessions(ctx context.Context) error {
+	_, err := c.grpc.DeleteAllAppSessions(ctx, &empty.Empty{})
+	return trail.FromGRPC(err)
+}
+
+// GenerateAppToken creates a JWT token with application access.
+func (c *Client) GenerateAppToken(ctx context.Context, req jwt.GenerateAppTokenRequest) (string, error) {
+	resp, err := c.grpc.GenerateAppToken(ctx, &proto.GenerateAppTokenRequest{
+		Username: req.Username,
+		Roles:    req.Roles,
+		URI:      req.URI,
+		Expires:  req.Expires,
+	})
+	if err != nil {
+		return "", trail.FromGRPC(err)
+	}
+
+	return resp.GetToken(), nil
+}
+
+// DeleteKubeService deletes a named kubernetes service.
+func (c *Client) DeleteKubeService(ctx context.Context, name string) error {
+	_, err := c.grpc.DeleteKubeService(ctx, &proto.DeleteKubeServiceRequest{
+		Name: name,
+	})
+	return trace.Wrap(err)
+}
+
+// DeleteAllKubeServices deletes all registered kubernetes services.
+func (c *Client) DeleteAllKubeServices(ctx context.Context) error {
+	_, err := c.grpc.DeleteAllKubeServices(ctx, &proto.DeleteAllKubeServicesRequest{})
+	return trace.Wrap(err)
+}
