+++ api/client/delegator.go
@@ -1,0 +1,44 @@
+/*
+Copyright 2020 Gravitational, Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package client
+
+import "context"
+
+type contextKey string
+
+const (
+	// ContextDelegator is a delegator for access requests set in the context
+	// of the request
+	ContextDelegator contextKey = "delegator"
+)
+
+// GetDelegator attempts to load the context value AccessRequestDelegator,
+// returning the empty string if no value was found.
+func GetDelegator(ctx context.Context) string {
+	delegator, ok := ctx.Value(ContextDelegator).(string)
+	if !ok {
+		return ""
+	}
+	return delegator
+}
+
+// WithDelegator creates a child context with the AccessRequestDelegator
+// value set.  Optionally used by AuthServer.SetAccessRequestState to log
+// a delegating identity.
+func WithDelegator(ctx context.Context, delegator string) context.Context {
+	return context.WithValue(ctx, ContextDelegator, delegator)
+}
