--- lib/auth/proto/proto.go
@@ -1,16 +1,0 @@
-package proto
-
-import "time"
-
-// Duration is a wrapper around duration to set up custom marshal/unmarshal
-type Duration time.Duration
-
-// Get returns time.Duration value
-func (d Duration) Get() time.Duration {
-	return time.Duration(d)
-}
-
-// Set sets time.Duration value
-func (d *Duration) Set(value time.Duration) {
-	*d = Duration(value)
-}
