+++ lib/auth/register.go
@@ -21,6 +21,7 @@ import (
 	"crypto/x509"
 
 	"github.com/gravitational/teleport"
+	"github.com/gravitational/teleport/api/client"
 	"github.com/gravitational/teleport/lib"
 	"github.com/gravitational/teleport/lib/defaults"
 	"github.com/gravitational/teleport/lib/services"
@@ -234,7 +235,7 @@ func insecureRegisterClient(params RegisterParams) (*Client, error) {
 		log.Infof("Joining remote cluster %v, validating connection with certificate on disk.", cert.Subject.CommonName)
 	}
 
-	client, err := NewTLSClient(ClientConfig{Addrs: params.Servers, TLS: tlsConfig})
+	client, err := NewClient(client.Config{Addrs: utils.NetAddrsToStrings(params.Servers), TLS: tlsConfig})
 	if err != nil {
 		return nil, trace.Wrap(err)
 	}
@@ -266,15 +267,15 @@ func pinRegisterClient(params RegisterParams) (*Client, error) {
 	// an attacker were to MITM this connection the CA pin will not match below.
 	tlsConfig := utils.TLSConfig(params.CipherSuites)
 	tlsConfig.InsecureSkipVerify = true
-	client, err := NewTLSClient(ClientConfig{Addrs: params.Servers, TLS: tlsConfig})
+	authClient, err := NewClient(client.Config{Addrs: utils.NetAddrsToStrings(params.Servers), TLS: tlsConfig})
 	if err != nil {
 		return nil, trace.Wrap(err)
 	}
-	defer client.Close()
+	defer authClient.Close()
 
 	// Fetch the root CA from the Auth Server. The NOP role has access to the
 	// GetClusterCACert endpoint.
-	localCA, err := client.GetClusterCACert()
+	localCA, err := authClient.GetClusterCACert()
 	if err != nil {
 		return nil, trace.Wrap(err)
 	}
@@ -300,12 +301,12 @@ func pinRegisterClient(params RegisterParams) (*Client, error) {
 	certPool.AddCert(tlsCA)
 	tlsConfig.RootCAs = certPool
 
-	client, err = NewTLSClient(ClientConfig{Addrs: params.Servers, TLS: tlsConfig})
+	authClient, err = NewClient(client.Config{Addrs: utils.NetAddrsToStrings(params.Servers), TLS: tlsConfig})
 	if err != nil {
 		return nil, trace.Wrap(err)
 	}
 
-	return client, nil
+	return authClient, nil
 }
 
 // ReRegisterParams specifies parameters for re-registering
