+++ lib/utils/addr.go
@@ -137,6 +137,15 @@ func (a *NetAddr) Set(s string) error {
 	return nil
 }
 
+// NetAddrsToStrings takes a list of netAddrs and returns a list of address strings.
+func NetAddrsToStrings(netAddrs []NetAddr) []string {
+	addrs := make([]string, len(netAddrs))
+	for i, addr := range netAddrs {
+		addrs[i] = addr.String()
+	}
+	return addrs
+}
+
 // ParseAddrs parses the provided slice of strings as a slice of NetAddr's.
 func ParseAddrs(addrs []string) (result []NetAddr, err error) {
 	for _, addr := range addrs {
